[TensorExpr] Explicitly cast to bool results of comparison ops in kernel.cpp.
Previously, we've been using operators <, >, ==, et al. and relied on
the dtype to be picked automatically. It led to a wrong dtype being
picked for the result, but that choice was overwritten by the type
explicitly specified in JIT IR, which we were lowering. Now we are
moving towards using shape inference instead of relying on all types
being specified in the IR, and that made this issue to immediately pop
up.

ghstack-source-id: 81f50e6bededf8c9f7d7c7e9c05e55fb4d39d084
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42201